Added to saveData and created new counterSet struct
This commit adds a counterSet struct which contains all the necessary
elements for the vector included in saveData. saveData now has a timer
(based off of the run model) which runs every 10 seconds. Every 10
seconds the vector will push a new element with the values of the
required variables. I have yet to created the totalBindingsCheck and the
totalAttemptsCheck, but will get on that.